chore(jit): fix circular dependencies
