+++ packages/jit/src/semantic-model.ts
@@ -1,6 +1,10 @@
 import { Immutable, IServiceLocator, PLATFORM } from '@aurelia/kernel';
 import { BindingMode, CustomAttributeResource, CustomElementResource, DOM, IBindableDescription, ICustomAttributeSource, IExpressionParser, IResourceDescriptions, ITemplateSource, TargetedInstruction } from '@aurelia/runtime';
-import { AttrSyntax, BindingCommandResource, Char, ElementSyntax, HydrateTemplateController, IAttributeParser, IBindingCommand, IElementParser, NodeType } from '.';
+import { AttrSyntax, IAttributeParser } from './attribute-parser';
+import { BindingCommandResource,  IBindingCommand } from './binding-command';
+import { Char } from './common';
+import { ElementSyntax, IElementParser, NodeType } from './element-parser';
+import { HydrateTemplateController } from './instructions';
 
 export class SemanticModel {
   public readonly isSemanticModel: true = true;
